+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/FirMetroErrors.kt
@@ -34,7 +34,7 @@ internal object FirMetroErrors : BaseDiagnosticRendererFactory() {
   val DEPENDENCY_GRAPH_ERROR by error1<String>(NAME_IDENTIFIER)
 
   // Inject constructor errors
-  val SUGGEST_CLASS_INJECTION_IF_NO_PARAMS by warning0(NAME_IDENTIFIER)
+  val SUGGEST_CLASS_INJECTION by warning0(NAME_IDENTIFIER)
 
   // Inject/assisted constructor errors
   val CANNOT_HAVE_MULTIPLE_INJECTED_CONSTRUCTORS by error0(NAME_IDENTIFIER)
@@ -98,8 +98,8 @@ internal object FirMetroErrors : BaseDiagnosticRendererFactory() {
 
       // Inject Constructor errors
       put(
-        SUGGEST_CLASS_INJECTION_IF_NO_PARAMS,
-        "There are no parameters on the @Inject-annotated constructor. Consider moving the annotation to the class instead.",
+        SUGGEST_CLASS_INJECTION,
+        "There is only one @Inject-annotated constructor. Consider moving the annotation to the class instead.",
       )
 
       // Inject/assisted Constructor errors