+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/fir/fir.kt
@@ -29,6 +29,7 @@ import org.jetbrains.kotlin.fir.declarations.FirClassLikeDeclaration
 import org.jetbrains.kotlin.fir.declarations.FirDeclarationOrigin
 import org.jetbrains.kotlin.fir.declarations.FirMemberDeclaration
 import org.jetbrains.kotlin.fir.declarations.FirTypeParameterRef
+import org.jetbrains.kotlin.fir.declarations.constructors
 import org.jetbrains.kotlin.fir.declarations.evaluateAs
 import org.jetbrains.kotlin.fir.declarations.findArgumentByName
 import org.jetbrains.kotlin.fir.declarations.getBooleanArgument
@@ -363,20 +364,24 @@ internal inline fun FirClassSymbol<*>.findInjectConstructor(
     0 -> null
     1 -> {
       constructorInjections[0].also {
-        if (it.isPrimary) {
+        val warnOnInjectAnnotationPlacement =
+          session.metroFirBuiltIns.options.warnOnInjectAnnotationPlacement
+        if (warnOnInjectAnnotationPlacement && constructors(session).size == 1) {
           val isAssisted =
             it.resolvedCompilerAnnotationsWithClassIds.isAnnotatedWithAny(
               session,
               session.classIds.assistedAnnotations,
             )
-          if (!isAssisted && it.valueParameterSymbols.isEmpty()) {
-            val inject =
-              it.resolvedCompilerAnnotationsWithClassIds
-                .annotationsIn(session, session.classIds.injectAnnotations)
-                .single()
-            if (KotlinTarget.CLASS in inject.getAllowedAnnotationTargets(session)) {
-              reporter.reportOn(inject.source, FirMetroErrors.SUGGEST_CLASS_INJECTION_IF_NO_PARAMS)
-            }
+          val inject =
+            it.resolvedCompilerAnnotationsWithClassIds
+              .annotationsIn(session, session.classIds.injectAnnotations)
+              .singleOrNull()
+          if (
+            !isAssisted &&
+              inject != null &&
+              KotlinTarget.CLASS in inject.getAllowedAnnotationTargets(session)
+          ) {
+            reporter.reportOn(inject.source, FirMetroErrors.SUGGEST_CLASS_INJECTION)
           }
         }
       }